% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flights_permuted.R
\docType{data}
\name{flights_permuted}
\alias{flights_permuted}
\title{Permuted nycflights13 data}
\format{
permuted flights data. Source data is the data from the nycflights13 package.
The data is included to test the diffdf feature of the package, which detects
differences between two datasets.
See commmit 41423e5510c4716a4b70fada1286c85dd698390a for the source script used
to generate the data.
}
\usage{
data(flights_permuted)
}
\description{
Permuted nycflights13 data
}
\keyword{datasets}
