% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_data_report.R
\name{render_data_report}
\alias{render_data_report}
\title{Title}
\usage{
render_data_report(
  df_input = NULL,
  save_report_to_disk = TRUE,
  df_input_old = NULL,
  save_rmd_dir = getwd(),
  save_html_dir = getwd(),
  include_skim = TRUE,
  include_diffdf = TRUE
)
}
\arguments{
\item{df_input}{data.frame or tibble. Input dataset to generate the summary report on.}

\item{save_report_to_disk}{boolean. TRUE to save the .Rmd file to disk at the specified
\code{save_rmd_dir} directory, FALSE to remove the generated .Rmd file at runtime.}

\item{df_input_old}{data.frame or tibble. Old input dataset to call diffdf::diffdf() on.}

\item{save_rmd_dir}{character. Path to save the .Rmd file to disk.}

\item{save_html_dir}{character. Path to save the .html output file to disk.}

\item{include_skim}{boolean. TRUE to include the data summary with skimr::skim() in the
report. FALSE to exclude.}

\item{include_diffdf}{boolean. TRUE to include the data diff with diffdf::diffdf() in the report.
FALSE to exclude.}
}
\value{
An output RMarkdown report with the data summary.
}
\description{
Title
}
