% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_data_report.R
\name{render_data_report}
\alias{render_data_report}
\title{Title}
\usage{
render_data_report(
  df_input = NULL,
  save_report_to_disk = TRUE,
  df_input_old = NULL,
  save_rmd_dir = getwd(),
  save_html_dir = getwd(),
  include_skim = TRUE,
  include_diffdf = TRUE
)
}
\arguments{
\item{df_input}{}

\item{save_report_to_disk}{}

\item{df_input_old}{}

\item{save_rmd_dir}{}

\item{save_html_dir}{}

\item{include_skim}{}

\item{include_diffdf}{}
}
\value{
An output RMarkdown report with the data summary.
}
\description{
Title
}
