% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_data_report.R
\name{render_data_report}
\alias{render_data_report}
\title{Title}
\usage{
render_data_report(
  df_input,
  df_input_old = NULL,
  save_rmd_dir = getwd(),
  save_html_dir = getwd(),
  include_skim = TRUE,
  include_diffdf = TRUE,
  output_format = "html"
)
}
\arguments{
\item{df_input}{data.frame or tibble. Input dataset to generate the summary report on.}

\item{df_input_old}{data.frame or tibble. Old input dataset to call diffdf::diffdf() on.}

\item{save_rmd_dir}{character. Path to save the .Rmd file to disk. Defaults to the current
working directory. If NULL, does not save the report to disk.}

\item{save_html_dir}{character. Path to save the .html output file to disk. Defaults to the current
working directory.}

\item{include_skim}{boolean. TRUE to include the data summary with skimr::skim() in the
report. FALSE to exclude.}

\item{include_diffdf}{boolean. TRUE to include the data diff with diffdf::diffdf() in the report.
FALSE to exclude. If df_input_old is not specified, automatically set to FALSE.}

\item{output_format}{character. Output format of the data report. Defaults to "html." So far, only "pdf" and "html"
are supported.}
}
\value{
An output RMarkdown report with the data summary.
}
\description{
Title
}
